Convert client to sending UTF-8 encoded JSON
The long and short of this is that pickle, for a networked application, is insecure. Official docs note that pickling isn't a good idea if security is a concern (which it obviously is.) To that end, everything is moving to hopefully harder to exploit JSON.

Keys and signatures are encoded in URL-safe base64. Updates are JSON objects (similar to before.)